+++ backend/src/app.js
@@ -2,6 +2,7 @@ const express = require('express');
 const morgan = require('morgan');
 const helmet = require('helmet');
 const cors = require('cors');
+const task = require('fake-todos');
 
 require('dotenv').config();
 
@@ -15,9 +16,12 @@ app.use(helmet());
 app.use(cors());
 app.use(express.json());
 
+
+
 app.get('/', (req, res) => {
+  var work = task(10);
   res.json({
-    message: '🦄🌈✨👋🌎🌍🌏✨🌈🦄'
+    message: work
   });
 });