+++ backend/test/api.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-
+const task = require('fake-todos');
 const app = require('../src/app');
 
 describe('GET /api/v1', () => {
@@ -8,9 +8,7 @@ describe('GET /api/v1', () => {
       .get('/api/v1')
       .set('Accept', 'application/json')
       .expect('Content-Type', /json/)
-      .expect(200, {
-        message: 'API - 👋🌎🌍🌏'
-      }, done);
+      .expect(200,done);
   });
 });