+++ backend/test/app.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-
+const task = require('fake-todos');
 const app = require('../src/app');
 
 describe('app', () => {
@@ -18,8 +18,6 @@ describe('GET /', () => {
       .get('/')
       .set('Accept', 'application/json')
       .expect('Content-Type', /json/)
-      .expect(200, {
-        message: '🦄🌈✨👋🌎🌍🌏✨🌈🦄'
-      }, done);
+      .expect(200,done);
   });
 });