+++ frontend/src/index.js
@@ -4,17 +4,11 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-const TODO_TASKS = [
-    {id: "todo-0", name:"Task 0", completed: true},
-    {id: "todo-1", name:"Task 1", completed: false},
-    {id: "todo-2", name:"Task 2", completed: false},
-];
-
 ReactDOM.render(
-  <React.StrictMode>
-    <App tasks={TODO_TASKS}/>
-  </React.StrictMode>,
-  document.getElementById('root')
+    <React.StrictMode>
+        <App/>
+    </React.StrictMode>,
+    document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function