+++ CHANGELOG.md
@@ -13,3 +13,10 @@
 
 - Initialised Pipeline  
 - Added docker image to pipeline so that universal url works
+
+# 0.4.0  
+
+- User can view, add and delete tasks in frontend
+- User can mark tasks as done from checkbox and task gets crossed out when marked as done
+- Made backend endpoints which can give 10, 5 or 1 randomly generated task
+- Frontend shows 3 randomly generated tasks from endpoint "http://21wsp8pw.course.tamk.cloud/api/v1/tasks/random"